@@ -1,39 +1,30 @@
 {
-  "name": "genericbot",
-  "version": "2.0.1",
-  "description": "GenericBot is an extremely simple Discord bot implemented with my boilerplate code",
-  "main": "src/index.js",
   "scripts": {
-    "start": "node src/index.js"
+    "start": "ts-node -r tsconfig-paths/register ./src/index.ts",
+    "lint": "eslint src/**/*.ts",
+    "dev": "ts-node-dev -r tsconfig-paths/register ./src/index.ts",
+    "debug": "node --inspect-brk=0.0.0.0:9229 -r ts-node/register -r tsconfig-paths/register ./src/index.ts"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/alex-taxiera/GenericBot.git"
-  },
-  "keywords": [
-    "generic",
-    "discord",
-    "bot",
-    "javascript"
-  ],
-  "author": "Alex Taxiera",
-  "license": "GPL-3.0",
-  "bugs": {
-    "url": "https://github.com/alex-taxiera/GenericBot/issues"
-  },
-  "homepage": "https://github.com/alex-taxiera/GenericBot#readme",
-  "dependencies": {
-    "dotenv": "^8.1.0",
-    "eris": "^0.10.1",
-    "eris-boiler": "github:alex-taxiera/eris-boiler",
-    "mysql": "github:mysqljs/mysql"
+  "engines": {
+    "node": ">=12.13.0"
   },
   "devDependencies": {
-    "eslint": "^6.5.1",
+    "@typescript-eslint/eslint-plugin": "^2.17.0",
+    "@typescript-eslint/parser": "^2.17.0",
+    "dotenv": "^8.2.0",
+    "eslint": "^6.8.0",
     "eslint-config-standard": "^14.1.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-node": "^10.0.0",
+    "eslint-plugin-import": "^2.20.0",
+    "eslint-plugin-node": "^11.0.0",
     "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1"
+    "eslint-plugin-standard": "^4.0.1",
+    "ts-node-dev": "^1.0.0-pre.44"
+  },
+  "dependencies": {
+    "@types/node": "^13.1.8",
+    "eris-boiler": "github:alex-taxiera/eris-boiler#release/3.0.0",
+    "ts-node": "^8.6.2",
+    "tsconfig-paths": "^3.9.0",
+    "typescript": "^3.7.5"
   }
 }
