@@ -1,12 +1,12 @@
-const { Command } = require('eris-boiler')
+import {
+  Command,
+  CommandResults
+} from 'eris-boiler'
 
-module.exports = new Command({
+export default new Command({
   name: 'raise',
   description: 'raise your blank',
-  options: {
-    deleteResponse: false
-  },
-  run: async ({ params }) => {
+  run: (_, { params }): CommandResults => {
     const fullParam = params.join(' ').toLowerCase()
 
     return `ヽ༼ຈل͜ຈ༽ﾉ raise your ${fullParam} ヽ༼ຈل͜ຈ༽ﾉ`
