@@ -1,25 +1,35 @@
-const { Command } = require('eris-boiler')
+import {
+  Command,
+  CommandResults
+} from 'eris-boiler'
 
-module.exports = new Command({
+function badChar (charCode: number): boolean {
+  return charCode > 122 || charCode < 97
+}
+
+function tooClose (indices: Array<number>, index: number): boolean {
+  return indices.includes(index) ||
+  indices.includes(index + 1) ||
+  indices.includes(index - 1)
+}
+
+export default new Command({
   name: 'spange',
   description: 'MaKe yOur tExT sPanGe',
-  options: {
-    deleteResponse: false
-  },
-  run: async ({ params }) => {
+  run: (_, { params }): CommandResults => {
     return params.map((param) => {
       param = param.toLowerCase()
-      if (param.split('').every((ch) => badChar(ch.charCodeAt()))) {
+      if (param.split('').every((ch) => badChar(ch.charCodeAt(0)))) {
         return param
       }
       const indices = []
 
-      let rng
-      let charCode = ''
+      let rng: number
+      let charCode: number
       for (let i = 0; i < (param.length * 0.33); i++) {
         do {
           rng = Math.floor(Math.random() * (param.length - 1 - 0 + 1)) + 0
-          charCode = param[rng].charCodeAt()
+          charCode = param[rng].charCodeAt(0)
         } while (badChar(charCode) || tooClose(indices, rng))
 
         indices.push(rng)
@@ -28,7 +38,7 @@ module.exports = new Command({
       let str = ''
       for (let i = 0; i < param.length; i++) {
         if (indices.includes(i)) {
-          str += String.fromCharCode(param[i].charCodeAt() - 32)
+          str += String.fromCharCode(param.charCodeAt(i) - 32)
         } else {
           str += param[i]
         }
@@ -37,13 +47,3 @@ module.exports = new Command({
     }).join(' ')
   }
 })
-
-function badChar (indices, charCode, rng) {
-  return charCode > 122 || charCode < 97
-}
-
-function tooClose (indices, index) {
-  return indices.includes(index) ||
-  indices.includes(index + 1) ||
-  indices.includes(index - 1)
-}
